powercap-rapl-sysfs: deprecate interface - use powercap-sysfs.h instead
By limiting zone depth to 2, this interface will break if Intel RAPL
extends its tree depth in the future.
The only other abstraction it provides is to mask the control type
parameter, which doesn't add much value.
Intel also added the intel-rapl-mmio control type, so even the control
type abstraction is not as useful as before.